Ace fix
Ace value subtraction has been fixed. Bet input validation has also been updated to include when a user inputs a number that is out of bounds.

Still to come: the rng for generating cards is very predictable. It would be nice if I could find a way to make it more random